helper: move to encrpytion directory
