+++ app/backup.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/pass-wall/passwall-server/api/login"
-	"github.com/pass-wall/passwall-server/helper"
 	"github.com/pass-wall/passwall-server/internal/database"
+	"github.com/pass-wall/passwall-server/internal/encryption"
 	"github.com/spf13/viper"
 )
 
@@ -58,6 +58,6 @@ func BackupData() error {
 		return err
 	}
 
-	helper.EncryptFile(backupPath, loginBytes.Bytes(), viper.GetString("server.passphrase"))
+	encryption.EncryptFile(backupPath, loginBytes.Bytes(), viper.GetString("server.passphrase"))
 	return nil
 }
